Added generate random, added data struct for results. Worked some more on date time.

+++ CppBasic/Assignment_1/Controller.cpp
@@ -13,9 +13,9 @@ void Controller::Run()
     }
 }
 
-float Controller::GetValidNumber()
+long long Controller::GetValidNumber()
 {
-    float num;
+    long long num;
     while (true)
     {
         std::cin >> num;
